+++ packages/mobile/react-native.config.js
@@ -1,3 +1,7 @@
 module.exports = {
-  assets: ['./assets/fonts', './assets/icons'],
-};
+  assets: ['./src/assets/fonts'],
+  resolver: {
+    assetExts: ['png', 'jpg', 'jpeg', 'gif', 'ttf'],
+    sourceExts: ['js', 'jsx', 'ts', 'tsx', 'svg'],
+  },
+}