+++ packages/mobile/src/components/JoinCommunity/JoinCommunity.component.tsx
@@ -116,7 +116,7 @@ export const JoinCommunity: FC<JoinCommunityProps> = ({
                 ]}
               />
             </View>
-            <View style={{ marginTop: 32 }}>
+            <View style={{ marginTop: 32 + 12 }}>
               <Button onPress={onPress} title={'Continue'} loading={loading} />
             </View>
           </KeyboardAvoidingView>